Move context extraction to new propagation file.
